Move repo_spec_from_find_results to repository_spec
Out of ansible_galaxy.install and into repository_spec
so it is easier to share.
